Note on speed issues:
I've thrown the Sonic 1 disassembly at OZ80 to see how quick it parses.
It's not good. There's a MAJOR slow down over increasing the array of
words; it takes too long to measure. Without that it takes 2.7 seconds.

I'm going to go back to the drawing board over my token format. A stream
of single bytes should be much faster and I set up a buffer to begin
with and increase the array in large chunks until the file is parsed.

Also: Added binary string syntax to SYNTAX.txt