Added Keyword parsing, changed purpose of main function
Since we're going directly from splitting words to tokens, we no longer need to hold an array of word strings. This saves memory and speed and is one less pass we need to make over the source code. A grammer check will be run over the tokens (instead of previously, over the source text words) which will be quicker to do.