Modified logical flow so we never have to subtract from TextPos
Cleaner flow and faster. I'm doing this so that we can easily capture the word begin/end offsets and eventually do away with all string concatenation entirely.

This has changed the number of words parsed, so there could be a bug, or there was a bug and I fixed it. I will have to verify the tokens being produced to ensure it's not missing things.

Also:
* Added paren / block tokens