Fixing bug with tokens being skipped
Changed `ScopeRoot` to loop internally rather than externally so that we can opt not to advance the token pointer in instances where it has already been done for us (reading expressions for example)

Also started working out label definitions with a CRC class to hash string names into indexes for quick look-up. I will need to take copies of the tokenStream's strings in the assembler so that we can concatenate properties to string-hashes.