Merged token numbers and token data numbers
This simplifies many things by having a single token for a Z80 instruction, instead of a "Z80" token, and then a data value for
particular instruction. Some slight complexity is added because checking if a particular token is a register / keyword / instruction &c. is not so immediately simple, but this was easily resolved by adding lookup tables to the TokenStream class.