Removed the string database from Assembler, and moved it into TokenStream
Strings were being stored in Assembler so as to pool together all strings from all TokenStream instances (source code files) so that this stuff is available without having to know which file it came from (i.e. cross-file references).

That's not really the right approach beacuse an individual string from a token is not the whole picture. A property such as ".loop" is meaningless on its own as it needs to be normalised to its parent, e.g. "::section:label.loop" and these names are stored in the label / variable databases accordingly.

So we'll be storing only label / variable &c. names in the Assembler class and not string fragments, which will stay in the TokenStream instances.