Redesign of some syntax
- Removed unimplemented keywords. I'm going to try stick to adding
things as I implement them to avoid confusion and now that I know the
design I'm going to be using
- Added constants to the design `DEF !CONSTANT $FF`, this fills in a
large missing part of the design
- Added Enums for autonumbering of constants
- Removed Variables entirely; I realised that constants serve the
purpose of variables in all purposes except within Macros, and I'm some
time away from implementing those. Plus also, the design of OZ80 is for
encapsulation of code/data and not linearly issue mode-changes. The
chunks of code/data that follow a statement should not be affected by
that statement!
- Z80 instructions now check if their value parameter is 8 or 16-bit
- Fixes and improvements to opcode building
- Massive overhaul and improvements to SYNTAX.txt -- trying to document
all new paging design and fill in the blanks
- Lots of small adjustments to the disassembly