MAJOR change to syntax
Changed parentheses for Z80 memory references "(IX+$FF)" to square
brackets "[IX+$FF]" and expression nesting to parentheses "3 + (1 + 2)"
instead of braces.

I knew this change would be necessary eventually, but I found the
collision point where the HELP parameter was trying to read the PROC
body as a nested expression.