Completely flattened the assembler
OZ80 is just one class now (plus the blu dependencies). This is a little crazy, but it will be much faster. The code is quite dirty at the moment as I need to clean up the integration of the File class directly into the main assembler.

Doing this also means that there's no global module and we don't have to fill the global space with a load of cache data