Redesign of error message / handling
Before I can reasonably focus on getting OZ80 to parse everything (rather than skipping stuff it doesn't yet understand), I need far more complete error handling. One part of this is to do away with a unique error number for every possibility and instead provide unique error messages instead of common ones. This will be expanded to include a copy of the line from the file pointing to the particular error location. A lot more work needs to be done on checking for errors in the parsing, particulary with Expressions / Values