Breakthrough milestone!
Six months ago I converted the WLADX assembly source to OZ80 -- a language that didn't exist except in theory, with no actual code capable of assembling it. I did this to get a feel for the Syntax design as I would be using it across a large code base.

Today, that theoretical language and assembly source code was assembled. 99% of it was correct from the beginning. Whilst OZ80 is incomplete and does not yet fill in the "indefinite" values (i.e. Label addresses can only be known after all code is parsed; a second assembly is needed to fill in these blanks), it has produced a binary that is in step with the original Sonic ROM. Every byte in the first 48 KB is in the correct place, even though many have the incorrect Value because the assembler simply isn't complete yet.

From the start of this project over a year ago until now, I have been working on an entirely theoretical product that I did not know if it would work, or if it was even possible to do.

With this commit, we have proof that this is a real prototype assembler implementing an entirely new programming language of my design!

* Added parsing of the SEGA 1 KB Slot pattern
* Filled in the remaining missing Sections on Sonic 1 assembly
* Extremely crude parsing of Text to bytes with `BYTE`
* Fix to Z80 assembly adding extra byte with RST instruction
* Fix to Z80 assembly ignoring 16-bit Value for unknown Label addresses
* Fix to off-by-one memory overflow in ROM Layout routine