The Naïveté Story
You know when you thought you were progressing well, but then you're suddenly hit with the realisation that you've overlooked something big and now loads of your work is redundent? I hit something big and had to do some soul-searching for a couple of days to work out how to proceed, if at all.

I'm treading well troden ground in the world outside of VB6, but I'm not formally trained, and within VB6 either there is no precedent, or what there is hopelessly inaccessible.

I overlooked the requirement of a virtual machine earlier on. Doing a second pass by walking the machine representation (token stream) of the source code files is very messy as a large portion of the work is definition and validation, which is not needed the second time around. Also, when we start adding turing-completeness like loops then things will get really slow doing two passes.

The answer is that rather than assembling from the token stream, we need to use the token stream to validate correctness of the OZ80 syntax and structure, but write out a virtual machine program to actually run the user's "program" which includes the Z80 portions.

So I will have to take two steps back to take one forward, ah well.

We have a new class "OZ80VM". I have designed a simple 'tape'-based virtual machine optimised for reading and writing files. A Tape is a binary array in memory that has a pointer/index (the Head) that can be moved around and data read and written. This universal concept allows the VM to handle files, data structures, arrays and more in a single, consistent manner. It's pretty neat actually.