Fixes to Type parsing
Sonic1 won't assemble ATM as we need to implement the `APPEND` keyword.
The main RAM block is currently defined twice, once in "ram.oz80" and
again in the sound driver. The idea is that `APPEND` can continue any
existing Chunk / Type / RAM.

TODO: We need to correctly measure the size of Types as this incorrectly
uses the address, which can be changed with can be assigned for each
property (e.g. enum) by the user