In Progress: Grammar validation & syntax tree
It's hideous -- don't look! So this is a complete overhaul of the file parsing. Instead of just tokenising words, we're now doing the grammar validation here, rather than during assembly. This will make things a lot faster whenever we introduce looping / macros / functions.

Much of the text parsing has been optimised too, so that we're faster even though we're doing more :)

The plan is to do away with the token stream and use an Abstract Syntax Tree which will remove the need to do so much validation during assembly. There's a huge amount of this portion of work remaining.