Comment-out the token-stream assembling...
All of the assembling will be rewritten (following the same pattern as
the token-stream assembling) using the Abstract Syntax Tree. For now,
source files are tokenised in one step, then converted and appended to
the syntax tree in another step (this will allow for file-caching of
token-streams in the future).

Some initial work is being done on pre-allocating Named Values (such as
Constants and Labels) so that these can be accessed quickly via IDs
rather than doing any array searching during assembly.